Support for multiple dropdowns + bugfix
This enables the dropdown feature to work (as long as foundation.dropdown.js is added to after foundation.js, refer to my recent merge request for foundation.dropdown.js)

I added the ability to enumerate through different dropdowns allowing the navbar to have multiple dropdowns.

This is done by using forloop.index and using that to create unique ID's for the UL's. Try it out!